requirements.txt
cassio
datasets
langchain 
openai 
tiktoken
pypdf2
python-dotenv
langchain_community 
langchain_openai